+++ components/equihash/src/verify.rs
@@ -116,7 +116,7 @@ impl fmt::Display for Kind {
     }
 }
 
-fn initialise_state(n: u32, k: u32, digest_len: u8) -> Blake2bState {
+pub(crate) fn initialise_state(n: u32, k: u32, digest_len: u8) -> Blake2bState {
     let mut personalization: Vec<u8> = Vec::from("ZcashPoW");
     personalization.write_all(&n.to_le_bytes()).unwrap();
     personalization.write_all(&k.to_le_bytes()).unwrap();